added a md file call module contribution guide to see how to add module or maybe prompt your ai
