+++ lib/features/common/remote/supabase_client.dart
@@ -1,0 +1,3 @@
+import 'package:supabase_flutter/supabase_flutter.dart';
+
+final SupabaseClient supabase = Supabase.instance.client;
