+++ lib/features/matchmaking/model/matchmaking_page.dart
@@ -1,0 +1,13 @@
+enum MatchmakingPage {
+  discover,
+  filters,
+  details,
+  create,
+  edit,
+  myTrips,
+  request,
+  sent,
+  manage,
+  applicant,
+  profile,
+}
