+++ lib/features/matchmaking/repository/matchmaking_repository.dart
@@ -1,0 +1,18 @@
+import 'package:flutter_riverpod/flutter_riverpod.dart';
+
+final matchmakingRepositoryProvider = Provider<MatchmakingRepository>(
+  (ref) => const MatchmakingRepository(),
+);
+
+class MatchmakingRepository {
+  const MatchmakingRepository();
+
+  List<String> get discoveryFilters => const [
+        'All',
+        'Adventure',
+        'Culture',
+        'Luxury',
+        'Nature',
+        'Foodie',
+      ];
+}
