+++ lib/features/profile/repository/profile_repository.dart
@@ -6,7 +6,7 @@ import 'package:shared_preferences/shared_preferences.dart';
 
 import '../../../constants/constants.dart';
 import '../../../features/profile/model/profile.dart';
-import '../../../main.dart';
+import '../../common/remote/supabase_client.dart';
 import '../../../utils/utils.dart';
 import '../../premium/model/premium_info.dart';
