+++ lib/routing/router.dart
@@ -6,7 +6,7 @@ import '../features/authentication/ui/otp_screen.dart';
 import '../features/authentication/ui/register_screen.dart';
 import '../features/authentication/ui/sign_in_screen.dart';
 import '../features/authentication/ui/welcome_screen.dart';
-import '../features/main/ui/main_screen.dart';
+import '../features/matchmaking/ui/matchmaking_shell_screen.dart';
 import '../features/onboarding/ui/onboarding_screen.dart';
 import '../features/onboarding/ui/splash_screen.dart';
 import '../features/premium/ui/premium_screen.dart';
@@ -111,7 +111,8 @@ final GoRouter router = GoRouter(
     ),
     GoRoute(
       path: Routes.main,
-      pageBuilder: (context, state) => state.slidePage(const MainScreen()),
+      pageBuilder: (context, state) =>
+          state.slidePage(const MatchmakingShellScreen()),
     ),
     GoRoute(
       path: Routes.accountInformation,
